IMPORTANT CHANGE!!!
this version change low level media decoder from wmplib to NAudio.
NAudio is a open source decoder and player for mp3 and wav,it is writen by C# and use very few resource to decode media file.
So I decide to use NAudio lib from this version.

this commit has done for play a song, but many UI designs must fix because of the change. There is a lot to do.